saving current state
